Fix namespace column logging: move parser to maf package
